Extract production code to separate files
